pyproject.toml
[project]
name = "twitter-mcp"
version = "0.1.0"
description = "Consolidated Twitter MCP server — Grok synthesis + X API v2"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.12.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
twitter-mcp = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["clients", "tools"]
include = ["server.py", "config.py"]

[tool.ruff]
line-length = 100
target-version = "py312"
